% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikertEZ.R
\name{export_summary_excel}
\alias{export_summary_excel}
\title{Export Summary to Excel}
\usage{
export_summary_excel(data, filename = "summary_report.xlsx", max_scale = 5)
}
\description{
Export Summary to Excel
}
