% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikertEZ.R
\name{import_likert_csv}
\alias{import_likert_csv}
\title{Import Likert data from CSV and convert factor columns to numeric}
\usage{
import_likert_csv(filepath)
}
\description{
Import Likert data from CSV and convert factor columns to numeric
}
