% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikertEZ.R
\name{summarize}
\alias{summarize}
\title{Summary for a single ordinal item (Likert-type)}
\usage{
summarize(responses, max_scale = 5, exact = TRUE, B = 10000, tidy = FALSE)
}
\arguments{
\item{responses}{Numeric vector}

\item{max_scale}{Max Likert scale value (default: 5)}

\item{exact}{Use Monte Carlo chi-sq when needed (default: TRUE)}

\item{B}{Number of replicates for simulation}

\item{tidy}{Return result as data.frame}
}
\value{
List or data.frame with stats
}
\description{
Summary for a single ordinal item (Likert-type)
}
