% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikertEZ.R
\name{summary_table_all}
\alias{summary_table_all}
\title{Create a tidy summary table of all items}
\usage{
summary_table_all(data, max_scale = 5, scale_labels = NULL, decimals = 2)
}
\arguments{
\item{data}{A data.frame of ordinal items}

\item{max_scale}{Max value on the scale}

\item{scale_labels}{Optional vector of labels for each scale point}

\item{decimals}{Number of decimal places for percentages (default = 2)}
}
\description{
Create a tidy summary table of all items
}
